Add support for picolibc TLS
Picolibc supports standard __thread variables and uses those
for internal per-thread values. For RTOS support, it exports
functions to initialize and switch TLS blocks at runtime.

This patch adds calls into picolibc using those functions, along with
allocating a TLS storage block for each thread. The size of that block
is not known until link time, so we have the application specify the
space and add a configASSERT check to verify that the space is
adequate.

Signed-off-by: Keith Packard <keithp@keithp.com>